Made Vec2f class use only floats after forgetting it used doubles. Very good performance increase. Is now memory bound

+++ include/vec2f.h
@@ -15,16 +15,16 @@ public:
 	union
 	{
 		float Element[2];
-		struct { double X, Y; };
+		struct { float X, Y; };
 	};
 
 public:
 	CUDA_CALLABLE_MEMBER Vec2f() {}
 
-	CUDA_CALLABLE_MEMBER Vec2f(double value)
+	CUDA_CALLABLE_MEMBER Vec2f(float value)
 		: X(value), Y(value) {}
 
-	CUDA_CALLABLE_MEMBER Vec2f(double x, double y)
+	CUDA_CALLABLE_MEMBER Vec2f(float x, float y)
 		: X(x), Y(y) {}
 
 	CUDA_CALLABLE_MEMBER Vec2f(const Vec2f &p_vec)
@@ -41,11 +41,11 @@ public:
 		return *this;
 	}
 
-	CUDA_CALLABLE_MEMBER inline Vec2f operator*(double val) const {
+	CUDA_CALLABLE_MEMBER inline Vec2f operator*(float val) const {
 		return Vec2f(val * X, val * Y);
 	}
 
-	CUDA_CALLABLE_MEMBER inline Vec2f operator/(double val) const 
+	CUDA_CALLABLE_MEMBER inline Vec2f operator/(float val) const
 	{
 		assert(val != 0.0);
 		return Vec2f(*this * (1.0f / val));
@@ -67,7 +67,7 @@ public:
 		return Vec2f(-X, -Y);
 	}
 
-	CUDA_CALLABLE_MEMBER inline Vec2f& operator*=(double val) {
+	CUDA_CALLABLE_MEMBER inline Vec2f& operator*=(float val) {
 		return *this = *this * val;
 	}
 
@@ -75,7 +75,7 @@ public:
 		return *this = *this * p_vec;
 	}
 
-	CUDA_CALLABLE_MEMBER inline Vec2f& operator/=(double val) {
+	CUDA_CALLABLE_MEMBER inline Vec2f& operator/=(float val) {
 		return *this = *this / val;
 	}
 
@@ -92,6 +92,6 @@ public:
 	}
 };
 
-CUDA_CALLABLE_MEMBER inline Vec2f operator*(double val, const Vec2f &p_vec) {
+CUDA_CALLABLE_MEMBER inline Vec2f operator*(float val, const Vec2f &p_vec) {
 	return Vec2f(val * p_vec.X, val * p_vec.Y);
 }
